requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
boltons==25.0.0
certifi==2026.1.4
colorama==0.4.6
face==24.0.0
frozenlist==1.8.0
glom==25.12.0
greenlet==3.3.0
idna==3.11
loguru==0.7.3
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
soupsieve==2.8.1
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
win32_setctime==1.2.0
yarl==1.22.0